+++ src/main/java/de/samply/db/repository/QueryRepository.java
@@ -1,0 +1,10 @@
+package de.samply.db.repository;
+
+import de.samply.db.model.Query;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface QueryRepository extends JpaRepository<Query, Long> {
+
+}