+++ src/main/resources/db/migration/V002__test_queries.sql
@@ -1,1 +1,2 @@
-INSERT INTO samply.query(query,format,label,description,contact_id, expiration_date) VALUES ('Patient', 'FHIR_QUERY', 'Test-1-20230110', 'Test 1', 'test-user', '2023-12-01');
+INSERT INTO samply.query(query, format, label, description, contact_id, created_at, expiration_date)
+VALUES ('Patient', 'FHIR_QUERY', 'Test-1-20230110', 'Test 1', 'test-user', now(), '2023-12-01');